+++ .gitignore
@@ -114,6 +114,9 @@ venv.bak/
 .spyderproject
 .spyproject
 
+# VS Code settings
+.vscode
+
 # Rope project settings
 .ropeproject
